Moved XML configuration element code to Spark.Web
- This code is not going to be relevant for .net core stuff (JSON configuration would be used instead)
- Breaking change the section type attribute will have to change to "Spark.Configuration.SparkSectionHandler, Spark.Web"
