--- src/Spark.Tests/App.config
@@ -1,36 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-  <configSections>
-    <section name="spark" type="Spark.Configuration.SparkSectionHandler, Spark"/>
-  </configSections>
-  <spark>
-    <compilation debug="true" nullBehaviour="Strict">
-      <assemblies>
-        <add assembly="System.Web, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a"/>
-      </assemblies>
-    </compilation>
-    <pages pageBaseType="Spark.Tests.Stubs.StubSparkView" automaticEncoding="false">
-      <namespaces>
-        <add namespace="System"/>
-      </namespaces>
-    </pages>
-  </spark>
-  <system.codedom>
-    <compilers>
-      <compiler language="vb;vbs;visualbasic;vbscript" extension=".vb" warningLevel="4" type="Microsoft.VisualBasic.VBCodeProvider, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-        <providerOption name="CompilerVersion" value="v4.0"/>
-        <providerOption name="OptionInfer" value="true"/>
-        <providerOption name="WarnAsError" value="false"/>
-      </compiler>
-    </compilers>
-    <!--<compilers>
-      <compiler language="c#;cs;csharp" extension=".cs" warningLevel="0" type="Microsoft.CSharp.CSharpCodeProvider, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-        <providerOption name="CompilerVersion" value="v4.0"/>
-        <providerOption name="WarnAsError" value="false"/>
-      </compiler>
-    </compilers>-->
-  </system.codedom>
-  <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
-  </startup>
-</configuration>
