+++ src/Spark.Web.Tests/App.config
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <configuration>
   <configSections>
-    <section name="spark" type="Spark.Configuration.SparkSectionHandler, Spark"/>
+    <section name="spark" type="Spark.Configuration.SparkSectionHandler, Spark.Web"/>
   </configSections>
   <spark>
     <compilation debug="true" nullBehaviour="Strict">
