+++ src/Spark.Web/Configuration/ResourcePathElement.cs
@@ -1,0 +1,42 @@
+// Copyright 2008-2009 Louis DeJardin - http://whereslou.com
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// 
+using System.Configuration;
+
+namespace Spark.Configuration
+{
+    public class ResourcePathElement : ConfigurationElement
+    {
+        [ConfigurationProperty("match")]
+        public string Match
+        {
+            get => (string)this["match"];
+            set => this["match"] = value;
+        }
+
+        [ConfigurationProperty("location")]
+        public string Location
+        {
+            get => (string)this["location"];
+            set => this["location"] = value;
+        }
+
+        [ConfigurationProperty("stop")]
+        public bool Stop
+        {
+            get => (bool) this["stop"];
+            set => this["stop"] = value;
+        }
+    }
+}
